Switch to an event queue
This change makes the EventListener merely listen for events and send
them to the GerritTrigger.  The GerritTrigger now has a background thread
called an EventProcessor.  When the GerritTrigger receives an event from
the EventListener, it sends the event to the GerritProcessor, which
maintains a queue.

In dynamic configuration mode, the queue is "paused" for reading until
the configuration has been loaded.  In non-dynamic configuration mode,
the queue is "unpaused" for reading immediately.

Between stop and start calls, the queue is maintained so that no events
are lost.
